requirements.txt
ipysketch~=0.0.1
scikit-image~=0.18.3
ipywidgets~=7.6.3
ipython~=7.26.0
setuptools~=57.4.0
Pillow~=8.3.1